move remove_defence_rows function to imputation helpers
