+++ tests/test_imputation/test_imputation_helpers.py
@@ -2,6 +2,7 @@ import pandas as pd
 import numpy as np
 from pandas import DataFrame as pandasDF
 from pandas._testing import assert_series_equal, assert_frame_equal
+from unittest.mock import patch
 import pytest
 
 from src.imputation.imputation_helpers import (
@@ -20,6 +21,7 @@ from src.imputation.imputation_helpers import (
     instance_fix,
     get_mult_604_mask,
     split_df_on_trim,
+    remove_defence_for_pnp,
 )
 
 
@@ -1017,3 +1019,75 @@ class TestSplitDfOnTrim:
         out_df1, out_df2 = split_df_on_trim(test_df, trim_bool_col)
         assert_frame_equal(out_df1.reset_index(drop=True), exp_df_trimmed.reset_index(drop=True))
         assert_frame_equal(out_df2.reset_index(drop=True), exp_df_not_trimmed.reset_index(drop=True))
+
+
+class TestRemoveDefenceForPNP():
+    """Tests for the remove_defence_for_pnp function."""
+    def create_input_df(self):
+        """Create an input dataframe for the test."""
+        input_columns = [
+            "reference",
+            "instance",
+            "200",
+            "legalstatus",
+            "statusencoded",
+            "postcodes_harmonised",
+        ]
+
+        data = [
+            [49900000404, 0, np.nan, "1", "210", "AB15 3GU"],
+            [49900000406, np.nan, np.nan, "2", "210", "BA1 5DA"],
+            [49900000409, 1, None, "1", "100", "CB1 3NF"],
+            [49900000510, 2, "D",  "7", "201", "BA1 5DA"],
+            [49900000510, 3, "D",  "7", "201", np.nan],
+            [49912758922, 3, "C", "1", "303", "DE72 3AU"],
+            [49900187320, 4, "C", "2", "304", "NP30 7ZZ"],
+            [49900184433, 1, "D", "7", "210", "CF10 BZZ"],
+            [49911791786, 1, np.nan, "4", "201", "CF10 BZZ"],
+            [49901183959, 4, "C", "1", "309", "SA50 5BE"],
+        ]
+
+        input_df = pandasDF(data=data, columns=input_columns)
+        return input_df
+
+    def create_exp_output_df(self):
+        """Create an output dataframe for the test."""
+        exp_output_columns = [
+            "reference",
+            "instance",
+            "200",
+            "legalstatus",
+            "statusencoded",
+            "postcodes_harmonised",
+        ]
+
+        data = [
+            [49900000404, 0, np.nan, "1", "210", "AB15 3GU"],
+            [49900000406, np.nan, np.nan, "2", "210", "BA1 5DA"],
+            [49900000409, 1, "C", "1", "100", "CB1 3NF"],
+            [49912758922, 3, "C", "1", "303", "DE72 3AU"],
+            [49900187320, 4, "C", "2", "304", "NP30 7ZZ"],
+            [49911791786, 1, "C", "4", "201", "CF10 BZZ"],
+            [49901183959, 4, "C", "1", "309", "SA50 5BE"],
+        ]
+        exp_output_df = pandasDF(data=data, columns=exp_output_columns)
+        return exp_output_df
+
+    def test_remove_defence_for_pnp(self):
+        """Test for the remove_defence_for_pnp function."""
+        exp_df = self.create_exp_output_df()
+        input_df = self.create_input_df()
+        result_df = remove_defence_for_pnp(input_df)
+        pd.testing.assert_frame_equal(
+            result_df.reset_index(drop=True), exp_df.reset_index(drop=True)
+        )
+
+        # Test if the function presents a list of  the defence refs to the info logger
+        #  before removing them
+        defence_rows = [49900000510, 49900184433]
+
+        with patch("src.imputation.imputation_helpers.ImputationHelpersLogger") as mock_logger:
+            remove_defence_for_pnp(input_df)
+            mock_logger.info.assert_called_with(
+                f"Defence rows found in PNP data: {defence_rows}"
+            )
